server_dev_

+++ controllers/admin.js
@@ -14,14 +14,15 @@ import fs from 'fs'
 import path from 'path'
 import { fileURLToPath } from 'url';
 import { compressAndSaveProductImage } from "../utils/compressImages.js";
-
+import NewsBlog from "../models/NewsModel.js";
+import CustomError from "../utils/ErrorClass.js";
 // Get current file directory
 const __filename = fileURLToPath(import.meta.url); 
 const __dirname = path.dirname(__filename);
 
 
 
-export const AdminSearchQuery=async(req,res)=>{
+export const AdminSearchQuery=async(req,res,next)=>{
 
     try{
 
@@ -123,9 +124,14 @@ export const AdminSearchQuery=async(req,res)=>{
 
        
 
-    }catch(error){
-        console.log(error)
-        res.status(501).json('Internal Server error')
+    }catch(err){
+       
+    
+        next(new CustomError(err.message, 500));
+      
+
+
+
     }
 }
 
@@ -137,7 +143,7 @@ export const AdminSearchQuery=async(req,res)=>{
 
 
 
-export const AdminLogin=async(req,res)=>{
+export const AdminLogin=async(req,res,next)=>{
   try{
     const {email,password} =req.body;
 
@@ -175,7 +181,9 @@ export const AdminLogin=async(req,res)=>{
       res.status(401).json('Unauthorized')
   }
   }catch(error){
-      console.log(error);
+    
+      next(new CustomError(err.message, 500));
+    
   }
 }
  
@@ -184,7 +192,7 @@ export const AdminLogin=async(req,res)=>{
 
 
 
-export const VerifyOTP=async(req,res)=>{
+export const VerifyOTP=async(req,res,next)=>{
   const {  otp } = req.body;
 
   try {
@@ -205,7 +213,7 @@ export const VerifyOTP=async(req,res)=>{
 
        const accessToken = generateAccessToken(user._id);
           const refreshToken = generateAdminRefreshToken(user._id);
-        console.log(refreshToken)
+       
           res.cookie('refreshToken', refreshToken, {
             httpOnly: true,
             path: '/', // Adjust the path as needed
@@ -216,9 +224,11 @@ export const VerifyOTP=async(req,res)=>{
 
     res.status(200).json({ accessToken:accessToken,permission:user.isAdmin });
   }
-  } catch (error) {
-    console.log(error)
-    res.status(500).json({ error: 'Error verifying OTP' });
+  } catch (err) {
+   
+    
+      next(new CustomError(err.message, 500));
+    
   }
 }
 
@@ -228,7 +238,7 @@ export const VerifyOTP=async(req,res)=>{
 
 
 
-export const AdminCsvFileHandling = (req, res) => {
+export const AdminProductsCsvFileHandling = (req, res,next) => {
   const results = [];
 
   if (!req.file) {
@@ -252,7 +262,7 @@ export const AdminCsvFileHandling = (req, res) => {
       try {
         for (const data of results) {
             try {
-              console.log(data);
+             
               const options = data.option.split(',').reduce((acc, opt, index) => {
                 const price = data.price.split(',')[index];
                 if (price) {
@@ -282,18 +292,21 @@ export const AdminCsvFileHandling = (req, res) => {
 
                 // Save the product to the database
                 await product.save();
-                console.log(`Product ${data.name} saved successfully.`);
-            } catch (error) {
-                console.error(`Error saving product ${data.name}: ${error.message}`);
-                // Optionally, you could store failed items in a separate array or log them for further inspection
+               
+            } catch (err) {
+             
+               
+              
+              
             }
         }
 
 
         res.status(200).send('CSV file processed and data saved successfully.');
-      } catch (error) {
-        console.error('Error saving data:', error);
-        res.status(500).send('Error processing CSV file.');
+      } catch (err) {
+        
+          next(new CustomError(err.message, 500));
+        
       } finally {
        
         fs.unlinkSync(req.file.path);
@@ -306,7 +319,96 @@ export const AdminCsvFileHandling = (req, res) => {
 
 
 
-export const AdminUploadImagesFolder=async (req,res)=>{
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+export const AdminBrandsCsvFileHandling = (req, res,next) => {
+  const results = [];
+
+  if (!req.file) {
+    return res.status(400).send('No file uploaded.');
+  }
+
+  // Read and parse the uploaded CSV file
+  fs.createReadStream(req.file.path)
+    .pipe(csvParser())
+    .on('data', (row) => {
+      // Preprocess the row to remove quotes and trim whitespace
+      const cleanedRow = {};
+      for (let key in row) {
+        // Remove any leading/trailing whitespace and extraneous quotes
+        cleanedRow[key.trim()] = row[key].trim().replace(/^"|"$/g, '');
+      }
+      // Push the cleaned row into the results array
+      results.push(cleanedRow);
+    })
+    .on('end', async () => {
+      try {
+        for (const data of results) {
+            try {
+             
+            
+                // Create a new product based on the CSV data
+                const brand = new Brand({
+                    name: data.name,
+                    brandLogoPath:data.brandImage
+                });
+
+                // Save the product to the database
+              const brandSaved=  await brand.save();
+              if(brandSaved){
+
+                const updateProducts=await Product.updateMany({
+                  $or: [
+                    { brand: brandSaved.name },          // Condition 1: Check for brand name
+                    { brandId: brandSaved._id }          // Condition 2: Check for brand ID
+                  ]
+                },{$set:{brandId:brandSaved._id,brand:brandSaved.name}})
+              }
+              
+               
+            } catch (err) { 
+             
+            
+               
+            }
+        }
+
+
+        res.status(200).send('CSV file processed and data saved successfully.');
+      } catch (err) {
+       
+        next(new CustomError(err.message, 500));
+      } finally {
+       
+        fs.unlinkSync(req.file.path);
+      }
+    });
+};
+
+
+
+
+
+
+
+export const AdminUploadProductImagesFolder=async (req,res,next)=>{
 
   try {
     const files = req.files;
@@ -322,7 +424,7 @@ export const AdminUploadImagesFolder=async (req,res)=>{
           const newFilename = `${Date.now()}-${file.originalname.replace(/ /g, '_').replace('.png', '_large.png')}`;
 
                 const newPath = path.join(productsDir, newFilename);
-                  console.log(newPath)
+                 
 
 
                 // Move the file to the new location with the new name
@@ -362,6 +464,98 @@ export const AdminUploadImagesFolder=async (req,res)=>{
         return res.status(400).json({ message: 'No files matched with any product images.' });
     }else{
       for (let file of matchedFiles){
+       
+        fs.unlinkSync(file.path);
+      }
+       
+    }
+
+    // Pass the matched files to the next middleware/controller
+    
+
+} catch (err) {
+ 
+  next(new CustomError(err.message, 500));
+}
+}
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+export const AdminUploadBrandImagesFolder=async (req,res)=>{
+
+  try {
+    const files = req.files;
+    const matchedFiles = [];
+    const brandsDir = path.join(__dirname, '../public/brands');
+ 
+    // Iterate over the uploaded files
+    for (let file of files) {
+       console.log(file)
+        const brand = await Brand.findOne({ brandLogoPath: file.originalname });
+
+        if (brand) {
+          const newFilename = `${Date.now()}-${file.originalname.replace(/ /g, '_')}`;
+
+                const newPath = path.join(brandsDir, newFilename);
+                  console.log(newPath)
+
+
+                // Move the file to the new location with the new name
+                fs.rename(file.path, newPath, async(err) => {
+                  if (err) {
+                    console.error(`Failed to rename and move file: ${err.message}`);
+                } else {
+                    
+                 brand.brandLogoPath=newFilename;
+                  
+                 await brand.save()
+                   
+
+
+
+
+
+                }
+                });
+
+
+                matchedFiles.push({ ...file, newFilename });
+            
+        } else {
+            // If no match, delete the file
+            fs.unlinkSync(file.path);
+        }
+    }
+
+    // If no files matched, send a response indicating no files were saved
+    if (matchedFiles.length === 0) {
+        return res.status(400).json({ message: 'No files matched with any product images.' });
+    }else{
+      for (let file of matchedFiles){
         console.log(matchedFiles)
         fs.unlinkSync(file.path);
       }
@@ -371,8 +565,154 @@ export const AdminUploadImagesFolder=async (req,res)=>{
     // Pass the matched files to the next middleware/controller
     
 
-} catch (error) {
-  console.log(error)
-    return res.status(500).json({ error: error.message });
+} catch (err) {
+  next(new CustomError(err.message, 500));
 }
 }
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+export const AdminNewsCsvFileHandling = (req, res) => {
+  const results = [];
+
+  if (!req.file) {
+    return res.status(400).send('No file uploaded.');
+  }
+
+  // Read and parse the uploaded CSV file
+  fs.createReadStream(req.file.path)
+    .pipe(csvParser())
+    .on('data', (row) => {
+      // Preprocess the row to remove quotes and trim whitespace
+      const cleanedRow = {};
+      for (let key in row) {
+        // Remove any leading/trailing whitespace and extraneous quotes
+        cleanedRow[key.trim()] = row[key].trim().replace(/^"|"$/g, '');
+      }
+      // Push the cleaned row into the results array
+      results.push(cleanedRow);
+    })
+    .on('end', async () => {
+      try {
+        for (const data of results) {
+            try {
+             
+           
+                // Create a new product based on the CSV data
+                const news = new NewsBlog({
+                    title: data.title,
+                    content:data.content,
+                    topic:data.topic,
+                    imageUrl:data.imageUrl
+                });
+
+                // Save the product to the database
+                await news.save();
+                console.log(`News ${data.title} saved successfully.`);
+            } catch (err) {
+              
+            }
+        }
+
+
+        res.status(200).send('CSV file processed and data saved successfully.');
+      } catch (err) {
+        next(new CustomError(err.message, 500));
+      } finally {
+       
+        fs.unlinkSync(req.file.path);
+      }
+    });
+};
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+export const AdminUploadNewsImagesFolder=async (req,res)=>{
+
+  try {
+    const files = req.files;
+    const matchedFiles = [];
+    const newsDir = path.join(__dirname, '../public/news');
+ 
+    // Iterate over the uploaded files
+    for (let file of files) {
+       console.log(file)
+        const news = await NewsBlog.findOne({ imageUrl: file.originalname });
+
+        if (news) {
+          const newFilename = `${Date.now()}-${file.originalname.replace(/ /g, '_')}`;
+
+                const newPath = path.join(newsDir, newFilename);
+                  console.log(newPath)
+
+
+                // Move the file to the new location with the new name
+                fs.rename(file.path, newPath, async(err) => {
+                  if (err) {
+                   
+                } else {
+                    
+                 news.imageUrl=newFilename;
+                  
+                 await news.save()
+                   
+
+
+
+
+
+                }
+                });
+
+
+                matchedFiles.push({ ...file, newFilename });
+            
+        } else {
+            // If no match, delete the file
+            fs.unlinkSync(file.path);
+        }
+    }
+
+    // If no files matched, send a response indicating no files were saved
+    if (matchedFiles.length === 0) {
+        return res.status(400).json({ message: 'No files matched with any News images.' });
+    }else{
+      for (let file of matchedFiles){
+        console.log(matchedFiles)
+        fs.unlinkSync(file.path);
+      }
+       
+    }
+
+    // Pass the matched files to the next middleware/controller
+    
+
+} catch (err) {
+  next(new CustomError(err.message, 500));
+}
+}
+
